requirements.txt
bcrypt==3.1.4
blinker==1.4
branca==0.2.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
Click==7.0
coverage==4.3.4
Django==2.1.2
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
Flask-WTF==0.14.2
folium==0.2.1
get==1.0.3
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Pillow==5.1.0
post==1.0.2
psycopg2==2.7.4
public==1.0.3
pycparser==2.19
pytz==2018.3
query-string==1.0.2
request==1.0.2
requests==2.19.1
selenium==3.11.0
six==1.11.0
SQLAlchemy==1.2.14
urllib3==1.22
virtualenv==15.1.0
Werkzeug==0.14.1
WTForms==2.2.1
